+++ pm1prob.h
@@ -11,7 +11,7 @@ extern "C" {
 #endif
 
 // Returns the probability of PM1(B1,B2) success for a Mersenne 2^exponent -1 already TF'ed to factoredUpTo.
-double pm1prob (double takeAwayBits, unsigned factoredUpTo, uint64_t B1, uint64_t B2);
+double pm1prob (double takeAwayBits, double factoredUpTo, uint64_t B1, uint64_t B2);
 
 // Dickman's "rho" function; rho(x) == F(1/x)
 double rho (double x);
